# docs/Architecture/openapi.yaml
openapi: 3.1.0
info:
  title: 闪卡 App 后端 API
  version: 2.5.0
  description: >
    闪卡 App v2.5 接口契约(机器可读版)。行为契约(状态机、幂等、FSRS 排程、
    错误码语义)见 structure-contract.md;需求依据 PRD v2.5(增量继承 v2.4~v2.1)。
    所有业务接口需携带 Authorization: Bearer(register/login 与探针豁免);写操作需携带
    Idempotency-Key。
    限流:写接口 60 req/min/user、IP 5 req/s,超限返回 429 + Retry-After(详见 structure-contract.md 1.6)。
    5xx(INTERNAL_ERROR / GENERATION_FAILED)由全局中间件统一处理,未逐操作声明。
    列表接口 MVP 暂不分页;幂等重放返回首次成功结果(2xx,含 204)。
    旧 /pdfs 兼容路径保留过渡期,内部委托项目接口同一业务语义(structure-contract 6.2)。
servers:
  - url: https://shanka.kbzz1.top
security:
  - BearerAuth: []
...
paths:
...
  /projects:
    post:
      tags: [projects]
      summary: 上传 PDF 建立学习项目(可选 name;上传成功即建立,PDF 异步解析)
      operationId: createProject
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
                  description: PDF 文件(≤100MB、≤1000 页)
                name:
                  type: string
                  description: 可选项目名;缺省取上传文件名去扩展名
      responses:
        '201':
          description: 项目已建立,PDF 进入解析
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningProject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '429':
          $ref: '#/components/responses/RateLimited'
    get:
      tags: [projects]
      summary: 当前用户项目列表(支持真实空态)
      operationId: listProjects
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                type: object
                required: [items]
                properties:
                  items:
                    type: array
                    items:
                      $ref: '#/components/schemas/LearningProject'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /projects/{project_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [projects]
      summary: 项目详情(含 file/chapters 摘要与派生计数;PARSING 时轮询)
      operationId: getProject
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningProject'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [projects]
      summary: 重命名(去首尾空白后 1~60 字符,可重名)
      operationId: renameProject
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name:
                  type: string
                  minLength: 1
                  maxLength: 60
      responses:
        '200':
          description: 更新后的项目
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningProject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      tags: [projects]
      summary: 删除项目(可原子取消活跃制卡任务;retain_decks 选择保留或删除全部项目牌组)
      operationId: deleteProject
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: retain_decks
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - name: abandon_pre_generation_tasks
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 同一事务内放弃 DRAFT/样卡阶段任务;GENERATING 仍返回冲突
        - name: cancel_active_tasks
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 二次确认后同一事务内取消全部活跃制卡任务(含 GENERATING)并继续删除
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{project_id}/deletion-preflight:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [projects]
      summary: 删除项目预检(只读;返回影响范围与活跃任务可执行动作)
      operationId: projectDeletionPreflight
      parameters:
        - name: retain_decks
          in: query
          required: false
          schema:
            type: boolean
            default: true
        - name: cancel_active_tasks
          in: query
          required: false
          schema:
            type: boolean
            default: false
          description: 预检是否展示“取消活跃任务并删除”动作;不修改数据
      responses:
        '200':
          description: 当前删除影响与阻塞任务
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeletionPreflight'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
  /projects/{project_id}/replace-pdf:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [projects]
      summary: 仅解析失败项目可替换并重新解析(原子替换 PDF)
      operationId: replaceProjectPdf
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        '200':
          description: 已替换并进入解析
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningProject'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{project_id}/chapters/{chapter_id}:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
      - $ref: '#/components/parameters/ChapterId'
    patch:
      tags: [projects]
      summary: 修改章节名称 / 起始页 / 结束页
      operationId: updateProjectChapter
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ChapterUpdateRequest'
      responses:
        '200':
          description: 已更新
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Chapter'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'
    delete:
      tags: [projects]
      summary: 删除章节(活跃任务保护;保留卡时 chapter_id 置空)
      operationId: deleteProjectChapter
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
        - name: delete_cards
          in: query
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '204':
          description: 已删除
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{project_id}/confirm-chapters:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    post:
      tags: [projects]
      summary: 确认目录,使项目进入 READY
      operationId: confirmProjectChapters
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      responses:
        '200':
          description: 已确认
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LearningProject'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '429':
          $ref: '#/components/responses/RateLimited'
  /projects/{project_id}/study-settings:
    parameters:
      - $ref: '#/components/parameters/ProjectId'
    get:
      tags: [projects]
      summary: 项目学习设置(新卡章节范围与未归属分组)
      operationId: getStudySettings
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStudySettings'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      tags: [projects]
      summary: 更新项目学习设置
      operationId: updateStudySettings
      parameters:
        - $ref: '#/components/parameters/IdempotencyKey'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectStudySettingsUpdateRequest'
      responses:
        '200':
          description: 更新后的设置
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectStudySettings'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'
        '429':
          $ref: '#/components/responses/RateLimited'

  # ---------- PDF ----------
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: opaque Bearer session token(注册/登录签发,服务端只存 SHA-256 摘要;决策 D-05)
  parameters:
    IdempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema:
        type: string
        format: uuid
      description: 幂等键,客户端生成;重复请求返回首次成功结果(见 structure-contract.md 1.3)
...
    ProjectId:
      name: project_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
    ChapterId:
      name: chapter_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
...
    PdfStatus:
      type: string
      enum: [PENDING, PARSING, PARSED, FAILED]
    TaskStatus:
      type: string
      enum: [DRAFT, SAMPLE_GENERATING, AWAITING_SAMPLE_CONFIRMATION, GENERATING, COMPLETED, FAILED, ABANDONED]
      description: V2.5 七态(structure-contract 4.1);历史 PENDING→DRAFT、RUNNING→GENERATING、CANCELLED→ABANDONED、PAUSED→FAILED(LEGACY_PAUSED_TASK)
    TaskInternalStage:
      type: string
      enum: [PLANNING, GENERATING, SCORING, PUBLISHING]
      description: 运行期内部阶段观测;不直接作为用户状态
...
    Error:
      type: object
      required: [error]
      description: 统一错误响应,外层 error 包装(契约 1.4)
      properties:
        error:
          type: object
          required: [code, message, localization_key]
          properties:
            code:
              type: string
              description: 稳定错误码(见 structure-contract.md 第 7 章)
            message:
              type: string
              description: 面向用户的补充信息,不含内部细节
            localization_key:
              type: string
            actions:
              type: [array, 'null']
              items:
                type: string
              description: 可执行动作提示,如 ABANDON_AND_RETRY / VIEW_TASKS / WAIT_FOR_TERMINAL
            details:
              type: [object, 'null']
              additionalProperties: true
...
    Chapter:
      type: object
      required: [chapter_id, name, start_page, end_page]
      properties:
        chapter_id:
          type: string
          format: uuid
        name:
          type: string
        start_page:
          type: integer
          minimum: 1
        end_page:
          type: integer
          minimum: 1
    ChapterUpdateRequest:
      type: object
      properties:
        name:
          type: string
          minLength: 1
          description: 章节名称,可修改
        start_page:
          type: integer
          minimum: 1
        end_page:
          type: integer
          minimum: 1
      description: 至少提供一个字段;未提供的字段保持不变(PRD 5.2)
    PdfFile:
      type: object
      required: [file_id, filename, size_bytes, status, created_at]
      properties:
        file_id:
          type: string
          format: uuid
        filename:
          type: string
        size_bytes:
          type: integer
        status:
          $ref: '#/components/schemas/PdfStatus'
        error_code:
          type: [string, 'null']
          description: PDF_PARSE_FAILED / PDF_TOC_MISSING
        chapters:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/Chapter'
        created_at:
          type: string
          format: date-time
...
    GenerationConfig:
      type: object
      required: [coverage_mode, difficulty_ratio]
      properties:
        coverage_mode:
          $ref: '#/components/schemas/CoverageMode'
        difficulty_ratio:
          $ref: '#/components/schemas/DifficultyRatio'
        custom_requirements:
          type: [string, 'null']
    TaskCursor:
      type: object
      required: [completed_batch_count]
      properties:
        completed_batch_count:
          type: integer
    Task:
      type: object
      required:
        - task_id
        - project_id
        - file_id
        - deck_id
        - status
        - selected_chapters
        - generation_config
        - generated_card_count
        - skipped_planning_group_count
        - resumable
        - created_at
        - updated_at
      properties:
        task_id:
          type: string
          format: uuid
        project_id:
          type: [string, 'null']
          format: uuid
          description: 归属项目;新任务必填,null 只兼容迁移前已失去 PDF 的终态历史任务(只读,不可重试)
        file_id:
          type: [string, 'null']
          format: uuid
          description: 新任务必填;删除 PDF 后置 null(任务保留)
        deck_id:
          type: [string, 'null']
          format: uuid
          description: 新任务必填(同项目);删除牌组后置 null(任务保留)
        retry_of_task_id:
          type: [string, 'null']
          format: uuid
          description: 失败重试关联(只指向同用户失败任务)
        operation_id:
          type: [string, 'null']
          format: uuid
          description: 持久化生成操作标识;用于跨请求/重启后的任务对账
        status:
          $ref: '#/components/schemas/TaskStatus'
        internal_stage:
          $ref: '#/components/schemas/TaskInternalStage'
          description: 运行期内部阶段观测,不直接作为用户状态
        selected_chapters:
          type: array
          items:
            $ref: '#/components/schemas/Chapter'
          description: 开始正式生成前冻结快照(契约 4.1)
        generation_config:
          $ref: '#/components/schemas/GenerationConfig'
        sample_cards:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/SampleCard'
          description: 持久化 1~3 张样卡;配置变化时清空
        sample_config_hash:
          type: [string, 'null']
        sample_confirmed_at:
          type: [string, 'null']
          format: date-time
        cursor:
          $ref: '#/components/schemas/TaskCursor'
        generated_card_count:
          type: integer
          description: 只统计已发布卡;失败任务为 0
        total_batch_count:
          type: [integer, 'null']
        completed_batch_count:
          type: [integer, 'null']
        completion_reason:
          type: [string, 'null']
          description: NO_GENERATION_UNITS(全组规划成功但 0 个合法单元的业务空结果,COMPLETED)
        skipped_planning_group_count:
          type: integer
          description: 部分规划组失败被跳过的组数(仅部分成功时 > 0)
        resumable:
          type: boolean
          description: 内部租约恢复判定(用户侧无 resume API)
        failure_stage:
          $ref: '#/components/schemas/FailureStage'
        error_code:
          type: [string, 'null']
        created_at:
          type: string
          format: date-time
        started_at:
          type: [string, 'null']
          format: date-time
        ended_at:
          type: [string, 'null']
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    ProjectStatus:
      type: string
      enum: [PARSING, PARSE_FAILED, AWAITING_CHAPTER_CONFIRMATION, READY]
    LearningProject:
      type: object
      required:
        - project_id
        - name
        - file
        - status
        - chapter_count
        - deck_count
        - task_count
        - created_at
        - updated_at
        - version
      properties:
        project_id:
          type: string
          format: uuid
        name:
          type: string
          description: 去首尾空白后 1~60 字符,可重名
        file:
          $ref: '#/components/schemas/PdfFile'
          description: 当前 PDF;列表响应可只返回摘要
        status:
          $ref: '#/components/schemas/ProjectStatus'
        chapter_count:
          type: integer
        deck_count:
          type: integer
        task_count:
          type: integer
        tasks:
          type: [array, 'null']
          items:
            $ref: '#/components/schemas/Task'
          description: 项目详情返回的任务快照;列表响应可省略
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        version:
          type: string
          description: 缓存刷新与并发检查
    ProjectStudySettings:
      type: object
      required: [selected_new_card_chapter_ids, include_unassigned, updated_at]
      properties:
        selected_new_card_chapter_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 只限制新卡;空数组表示暂无新卡范围
        include_unassigned:
          type: boolean
          description: 是否包含 chapter_id=null 的新卡
        updated_at:
          type: string
          format: date-time
    ProjectStudySettingsUpdateRequest:
      type: object
      properties:
        selected_new_card_chapter_ids:
          type: array
          items:
            type: string
            format: uuid
        include_unassigned:
          type: boolean
    DeletionTaskBlocker:
      type: object
      required: [task_id, status, can_abandon, can_cancel, allowed_actions]
      properties:
        task_id:
          type: string
          format: uuid
        status:
          $ref: '#/components/schemas/TaskStatus'
        internal_stage:
          type: [string, 'null']
          enum: [PLANNING, GENERATING, SCORING, PUBLISHING, null]
        project_id:
          type: [string, 'null']
          format: uuid
        deck_id:
          type: [string, 'null']
          format: uuid
        can_abandon:
          type: boolean
        can_cancel:
          type: boolean
          description: 是否可随资源删除一并取消
        allowed_actions:
          type: array
          items:
            type: string
            enum: [ABANDON_AND_RETRY, CANCEL_AND_DELETE, WAIT_FOR_TERMINAL, VIEW_TASKS]
    DeletionPreflight:
      type: object
      required: [resource_type, resource_id, can_delete, blockers, abandonable_task_ids, has_uncancellable_tasks, cancelable_task_ids, can_cancel, actions, impact]
      properties:
        resource_type:
          type: string
          enum: [PROJECT, DECK]
        resource_id:
          type: string
          format: uuid
        can_delete:
          type: boolean
        blockers:
          type: array
          items:
            $ref: '#/components/schemas/DeletionTaskBlocker'
        abandonable_task_ids:
          type: array
          items:
            type: string
            format: uuid
        has_uncancellable_tasks:
          type: boolean
        cancelable_task_ids:
          type: array
          items:
            type: string
            format: uuid
          description: 可随资源一并取消的活跃制卡任务
        can_cancel:
          type: boolean
          description: 是否允许采用取消任务后删除的决策
        actions:
          type: array
          items:
            type: string
            enum: [ABANDON_AND_RETRY, CANCEL_AND_DELETE, WAIT_FOR_TERMINAL, VIEW_TASKS]
        impact:
          type: object
          additionalProperties: true
...
  responses:
    BadRequest:
      description: 请求非法
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: >-
        受保护接口:缺失、非法、撤销或过期的 Bearer 会话(AUTH_REQUIRED / AUTH_INVALID,携带 WWW-Authenticate: Bearer);
        登录接口:凭据错误(INVALID_CREDENTIALS)
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: 资源不存在
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Conflict:
      description: 状态冲突或幂等内容不一致
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    RateLimited:
      description: 限流
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'